refactor(about): refactor about file
refactors the about file and downScroll file.
