+++ src/components/About/Variants.js
@@ -1,0 +1,41 @@
+export const aboutVariant = {
+  hidden: {
+    y: "20vh",
+    opacity: 0,
+  },
+  visible: {
+    y: 0,
+    opacity: 1,
+    transition: {
+      duration: 0.8,
+    },
+  },
+};
+
+export const leftAniVariant = {
+  hidden: {
+    x: "-20vh",
+    opacity: 0,
+  },
+  visible: {
+    x: 0,
+    opacity: 1,
+    transition: {
+      duration: 0.9,
+    },
+  },
+};
+
+export const rightAniVariant = {
+  hidden: {
+    x: "20vh",
+    opacity: 0,
+  },
+  visible: {
+    x: 0,
+    opacity: 1,
+    transition: {
+      duration: 0.7,
+    },
+  },
+};
