+++ src/components/Home/Home.js
@@ -1,6 +1,6 @@
 import { StyledHome } from "./styles";
 import heroIcon from "../../assets/heroIcon.svg";
-import ScrollDown from "../../components/ScrollDown";
+import ScrollDown from "../ScrollDown/ScrollDown";
 import { motion } from "framer-motion";
 import { greetingVariant, heroVariant } from "./Variants";
