--- src/components/styles/ScrollDown.styled.js
@@ -1,18 +1,0 @@
-import styled from "styled-components";
-
-export const StyledScroll = styled.div`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  p {
-    font-size: 20px;
-    color: #fefefe;
-  }
-
-  img {
-    margin-left: 10px;
-    height: 30px;
-    border-radius: 50%;
-  }
-`;
